Created addFoodScreen and styles directories
Created the logic and UI to add a new food item to the database
through the UI. The UI accepts the name of the food item and its
cost and validates the entries and outputs the entry added to the
database into the terminal. Created directories for each front-end
UI screens called screens and created directories to store UI styles
in the directory called styles to modularize the project code. The
app functions as its supposed to.