Created orderForm and savedOrderList screens
Edited crudOperations.dart to input 20 sample food values into the
database when the code runs. this function works. Added code for
two screens that allow the user to input a cost value and select
food items in their list. The 2nd screen shows the log of that.
It shows the list of all options selected with the date.